requirements.txt
blinker==1.6.2
click==8.1.7
colorama==0.4.6
Deprecated==1.2.14
Flask==2.3.3
Flask-JWT-Extended==4.5.2
Flask-Limiter==3.5.0
importlib-resources==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
limits==3.6.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
ordered-set==4.1.0
packaging==23.1
Pillow==10.0.0
Pygments==2.16.1
PyJWT==2.8.0
rich==13.5.2
typing_extensions==4.7.1
Werkzeug==2.3.7
wrapt==1.15.0
